requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-limiter==3.7.0
python-dotenv==1.0.1
gunicorn==22.0.0
google-genai==0.8.0
pytest==8.2.0
pytest-flask==1.3.0
pytest-cov==5.0.0
chartjs-python==0.1.6
